updated infosheet file structure and naming. also edited remaining eslint errors
