+++ src/pages/Help/InformationSheetDictionary/InformationSheetDictionary.js
@@ -8,11 +8,11 @@ import { Info } from '@mui/icons-material';
 import React, {
   useContext, useEffect, useState,
 } from 'react';
-import z4Dict from '../../shared/json/zone4/data-dictionary.json';
-import z5Dict from '../../shared/json/zone5/data-dictionary.json';
-import z6Dict from '../../shared/json/zone6/data-dictionary.json';
-import z7Dict from '../../shared/json/zone7/data-dictionary.json';
-import { Context } from '../../store/Store';
+import z4Dict from '../../../shared/json/zone4/data-dictionary.json';
+import z5Dict from '../../../shared/json/zone5/data-dictionary.json';
+import z6Dict from '../../../shared/json/zone6/data-dictionary.json';
+import z7Dict from '../../../shared/json/zone7/data-dictionary.json';
+import { Context } from '../../../store/Store';
 import DictionaryContent from './DictionaryContent';
 
 const InformationSheetDictionary = ({ zone, from }) => {
