+++ src/pages/InformationSheetContent/informationSheet.styles.js
